fix: breaking bug on localhost
