+++ client/src/utils/statsContract.js
@@ -1,57 +1,45 @@
-
-import { getTruffleContract, getNetworkNamefromId } from './ethutil';
+import { getTruffleContract } from './ethutil';
+import { getDeployData } from './deploycontract';
 import Web3 from 'web3';
 
 export const getLevelsSolvedByPlayer = async (playerAddress, networkId) => {
     if (!(playerAddress || networkId)) {
         return;
     }
-    const networkName = getNetworkNamefromId(networkId)
-    if (!networkName) { 
-        return
-    }
-    const levelAddresses = getLevelAddressesInNetwork(networkName)
-    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkName)
+    const levelAddresses = getLevelAddressesInNetwork(networkId)
+    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkId)
     const statsContract = await getStatsContract(proxyStatsAddress, playerAddress)
     const listOfLevelsSolved = await getListOfLevelsSolvedByPlayer(statsContract, levelAddresses, playerAddress, 5)
     return listOfLevelsSolved
 }
 
-export const getPercentageOfLevelsSolvedByPlayer = async (playerAddress, networkId) => { 
+export const getPercentageOfLevelsSolvedByPlayer = async (playerAddress, networkId) => {
     if (!(playerAddress || networkId)) {
         return
     }
-    const networkName = getNetworkNamefromId(networkId)
-    if (!networkName) { 
-        return
-    }
-    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkName)
+    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkId)
     const statsContract = await getStatsContract(proxyStatsAddress, playerAddress)
     return await getPercentageOfLevelsSolved(statsContract, playerAddress)
-     
+
 }
 
-export const checkIfPlayerExist = async (playerAddress, networkId) => { 
+export const checkIfPlayerExist = async (playerAddress, networkId) => {
     if (!(playerAddress || networkId)) {
         return
     }
-    const networkName = getNetworkNamefromId(networkId)
-    if (!networkName) { 
-        return
-    }
-    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkName)
+    const proxyStatsAddress = getProxyStatsContractAddressInNetwork(networkId)
     const statsContract = await getStatsContract(proxyStatsAddress, playerAddress)
     return await statsContract.doesPlayerExist(playerAddress)
 }
 
-const getPercentageOfLevelsSolved = async (statsContract, playerAddress) => { 
+const getPercentageOfLevelsSolved = async (statsContract, playerAddress) => {
     const response = await statsContract.getPercentageOfLevelsCompleted(playerAddress)
     const roundedPercentage = (Web3.utils.fromWei(response.toString()) * 100).toFixed(2)
     return roundedPercentage
 }
 
 
-const getListOfLevelsSolvedByPlayer = async (statsContract, levelAddresses, playerAddress, noOfQueriesAtAtime) => { 
+const getListOfLevelsSolvedByPlayer = async (statsContract, levelAddresses, playerAddress, noOfQueriesAtAtime) => {
     const isLevelSolvedRequestsList = getIsLevelSolvedRequestsList(statsContract, levelAddresses, playerAddress)
     const isLevelSolvedRequestsListChunked = chunk(isLevelSolvedRequestsList, noOfQueriesAtAtime)
     let isLevelSolvedList = []
@@ -62,28 +50,28 @@ const getListOfLevelsSolvedByPlayer = async (statsContract, levelAddresses, play
     return getListOfSolvedLevels(isLevelSolvedList)
 }
 
-const getListOfSolvedLevels = (isLevelSolvedList) => { 
+const getListOfSolvedLevels = (isLevelSolvedList) => {
     const listOfSolvedLevels = isLevelSolvedList
-        .filter(item=>item.isSolved)
-        .map(item=>item.levelAddress)
+        .filter(item => item.isSolved)
+        .map(item => item.levelAddress)
     return listOfSolvedLevels
 }
 
-const chunk = (list, chunkSize) => { 
+const chunk = (list, chunkSize) => {
     let chunkedList = [...Array(Math.ceil(list.length / chunkSize))].map(_ => list.splice(0, chunkSize))
     return chunkedList
 }
 
-const getIsLevelSolvedRequestsList = (statsContract, levelAddresses, playerAddress) => { 
+const getIsLevelSolvedRequestsList = (statsContract, levelAddresses, playerAddress) => {
     const levelNos = Object.keys(levelAddresses);
     const listOfIsLevelSolvedRequests = []
-    for (let levelNo of levelNos) { 
+    for (let levelNo of levelNos) {
         listOfIsLevelSolvedRequests.push(isLevelSolvedByPlayer(statsContract, levelAddresses[levelNo], playerAddress))
     }
     return listOfIsLevelSolvedRequests;
 }
 
-const isLevelSolvedByPlayer = async (statsContract, levelAddress, playerAddress) => { 
+const isLevelSolvedByPlayer = async (statsContract, levelAddress, playerAddress) => {
     const isSolved = await statsContract.isLevelCompleted(playerAddress, levelAddress)
     return {
         levelAddress,
@@ -91,24 +79,24 @@ const isLevelSolvedByPlayer = async (statsContract, levelAddress, playerAddress)
     };
 }
 
-const getLevelAddressesInNetwork = (networkName) => {
-    const deployedNetworkData = require(`../gamedata/deploy.${networkName}.json`);
+const getLevelAddressesInNetwork = (networkId) => {
+    const deployedNetworkData = getDeployData(networkId);
     const keys = Object.keys(deployedNetworkData);
     const levelAddresses = {}
-    for (let key of keys) { 
-        if (!isNaN(Number(key))) { 
+    for (let key of keys) {
+        if (!isNaN(Number(key))) {
             levelAddresses[key] = deployedNetworkData[key];
         }
     }
     return levelAddresses;
 }
 
-const getProxyStatsContractAddressInNetwork = (networkName) => {
-    const deployedNetworkData = require(`../gamedata/deploy.${networkName}.json`);
+const getProxyStatsContractAddressInNetwork = (networkId) => {
+    const deployedNetworkData = getDeployData(networkId);
     return deployedNetworkData['proxyStats'];
 }
 
-const getStatsContract = async (proxyStatsAddress, playerAddress) => { 
+const getStatsContract = async (proxyStatsAddress, playerAddress) => {
     const statsABI = require("contracts/build/contracts/metrics/Statistics.sol/Statistics.json");
     const statsContract = getTruffleContract(statsABI, { from: playerAddress });
     const statsContractInstance = await statsContract.at(proxyStatsAddress)
